Move IQXJsonEncoder tests to test_serialization file
